feat: supports double-click editing of marker text
