+++ packages/vchart-editor/src/elements/chart/editor-elements/marker/base.ts
@@ -1,9 +1,12 @@
 import type { IChartModel } from './../../interface';
-import type { IGraphic, IGroup, INode } from '@visactor/vrender-core';
+import type { IText } from '@visactor/vrender-core';
+import { type IGraphic, type IGroup, type INode } from '@visactor/vrender-core';
 import type { IEditorElement } from '../../../../core/interface';
 import { BaseEditorElement, CommonChartEditorElement } from '../base-editor-element';
 import type { EventParams, ICartesianSeries, IComponent } from '@visactor/vchart';
 import { MarkerTypeEnum } from '../../interface';
+import { setupSimpleTextEditor } from '../../utils/text';
+import { get } from '@visactor/vutils';
 
 export abstract class BaseMarkerEditor<T extends IComponent, D> extends BaseEditorElement {
   readonly type: string = 'marker';
@@ -11,12 +14,14 @@ export abstract class BaseMarkerEditor<T extends IComponent, D> extends BaseEdit
   protected _model: T; // vchart 组件模型实例
   protected _element: D; // 组件模型实例对应的 vrender 组件
   protected _modelId: string | number; // // vchart 组件模型实例 id
+  protected _spec: any; // 组件 spec
 
   protected _editComponent: IGroup; // 标注编辑元素
 
   protected abstract _getEnableMarkerTypes(): string[];
   protected abstract _createEditorGraphic(el: IEditorElement, e: PointerEvent): IGraphic;
   protected abstract _handlePointerDown(e: EventParams): void;
+  protected abstract _onTextChange(expression: string): void;
   protected abstract _setCursor(e: EventParams): void;
   protected _handlePointerUp(e: EventParams): void {
     this.endEditor();
@@ -24,10 +29,12 @@ export abstract class BaseMarkerEditor<T extends IComponent, D> extends BaseEdit
 
   initWithVChart(): void {
     const vchart = this._chart.vchart;
+
     vchart.on('pointermove', { level: 'model', type: this.type }, this._onHover);
     vchart.on('pointerdown', { level: 'model', type: this.type }, this._onDown);
     vchart.on('pointerup', { level: 'model', type: this.type }, this._onUp);
     vchart.on('pointerleave', { level: 'model', type: this.type }, this._onLeave);
+    vchart.on('dblclick', { level: 'model', type: this.type }, this._onDblclick);
   }
 
   private _checkEventEnable(e: EventParams) {
@@ -45,8 +52,12 @@ export abstract class BaseMarkerEditor<T extends IComponent, D> extends BaseEdit
 
     const el = this._getEditorElement(e);
     this.showOverGraphic(el, el?.id + `${this._layer.id}`, e.event as PointerEvent);
-    // this._modelId = el.model.userId;
-    this._setCursor(e);
+
+    if (get(e, 'event.target.name') === 'tag-text') {
+      this._chart.option.editorEvent.setCursor('pointer');
+    } else {
+      this._setCursor(e);
+    }
   };
 
   protected _onDown = (e: EventParams) => {
@@ -56,6 +67,7 @@ export abstract class BaseMarkerEditor<T extends IComponent, D> extends BaseEdit
     this._element = (<T>e.model).getVRenderComponents()[0] as unknown as D;
     this._model = e.model as T;
     this._modelId = e.model.userId;
+    this._spec = e.model.getSpec();
 
     this._handlePointerDown(e);
   };
@@ -77,6 +89,36 @@ export abstract class BaseMarkerEditor<T extends IComponent, D> extends BaseEdit
     }
     // 恢复 cursor
     this._chart.option.editorEvent.setCursorSyncToTriggerLayer();
+  };
+
+  protected _onDblclick = (e: EventParams) => {
+    if (!this._checkEventEnable(e)) {
+      return;
+    }
+
+    if (get(e, 'event.target.name') !== 'tag-text') {
+      return;
+    }
+    const vchart = this._chart.vchart;
+    const text = e.event.target as IText;
+    setupSimpleTextEditor({
+      text: text,
+      container: vchart.getContainer(),
+      panelStyle: {
+        padding: { left: 4, right: 4, top: 4, bottom: 4 },
+        lineWidth: 2
+      },
+      defaultFontFamily: vchart.getCurrentTheme().fontFamily,
+      expression: this._spec.expression,
+      needExpression: true,
+      change: (expression: string) => {
+        if (expression === this._spec.expression || (this._spec.expression === undefined && expression === '##')) {
+          return;
+        }
+
+        this._onTextChange(expression);
+      }
+    });
   };
 
   protected _getEditorElement(eventParams: EventParams): IEditorElement {
@@ -155,10 +197,7 @@ export abstract class BaseMarkerEditor<T extends IComponent, D> extends BaseEdit
     ];
     root.getChildren().forEach((child: INode) => {
       if (marks.includes(child.name)) {
-        (child as IGroup).setAttributes({
-          pickable: false,
-          childrenPickable: false
-        });
+        this._setMarkerShapePickable(child as IGraphic, false);
       }
     });
   }
@@ -175,19 +214,27 @@ export abstract class BaseMarkerEditor<T extends IComponent, D> extends BaseEdit
       MarkerTypeEnum.verticalArea,
       MarkerTypeEnum.horizontalArea
     ];
-    root.getChildren().forEach((child: INode) => {
+    root.getChildren().forEach(child => {
       if (marks.includes(child.name)) {
-        (child as IGroup).setAttributes({
-          pickable: true,
-          childrenPickable: true
-        });
+        // 只开启非 group 元素的拾取
+        this._setMarkerShapePickable(child as IGraphic, true);
       }
     });
   }
 
+  private _setMarkerShapePickable(node: IGraphic, pickable: boolean) {
+    if (node.isContainer) {
+      node.getChildren().forEach(child => {
+        this._setMarkerShapePickable(child as IGraphic, pickable);
+      });
+    } else {
+      node.setAttribute('pickable', pickable);
+    }
+  }
+
   protected _updateAndSave(spec: any, type: 'markLine' | 'markArea') {
     // 更新
-    this._chart.specProcess.updateElementAttribute(this._currentEl.model, {
+    this._chart.specProcess.updateElementAttribute(this._currentEl?.model, {
       [type]: {
         spec
       }
@@ -195,14 +242,16 @@ export abstract class BaseMarkerEditor<T extends IComponent, D> extends BaseEdit
     this._chart.reRenderWithUpdateSpec();
 
     // 更新更新后的标注的 bounds
-    const currentMark = this._chart.vchart.getStage().getElementById(spec.id) as IGraphic;
-    const markerBounds = currentMark.AABBBounds;
-    this._currentEl.updateRect({
-      x: markerBounds.x1,
-      y: markerBounds.y1,
-      width: markerBounds.width(),
-      height: markerBounds.height()
-    });
+    if (this._currentEl) {
+      const currentMark = this._chart.vchart.getStage().getElementById(spec.id) as IGraphic;
+      const markerBounds = currentMark.AABBBounds;
+      this._currentEl.updateRect({
+        x: markerBounds.x1,
+        y: markerBounds.y1,
+        width: markerBounds.width(),
+        height: markerBounds.height()
+      });
+    }
 
     this._controller.editorEnd();
   }
@@ -229,6 +278,10 @@ export abstract class BaseMarkerEditor<T extends IComponent, D> extends BaseEdit
 
     vchart.off('pointermove', this._onHover);
     vchart.off('pointerdown', this._onDown);
+    // vchart.off('pointerenter', this._onEnter);
+    vchart.off('pointerleave', this._onLeave);
+    vchart.off('dblclick', this._onDblclick);
+
     super.release();
   }
 }
