+++ packages/vchart-editor/src/elements/chart/utils/dom.ts
@@ -1,0 +1,10 @@
+export default function modifyCSS(dom: HTMLElement, css: { [key: string]: any }): HTMLElement {
+  if (dom) {
+    for (const key in css) {
+      if (css.hasOwnProperty(key)) {
+        dom.style[key] = css[key];
+      }
+    }
+  }
+  return dom;
+}
